requirements.txt
openai>=1.40.0
google-cloud-vision>=3.7.0
pdf2image>=1.17.0
Pillow>=10.0.0
pandas>=2.2.0
